+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0
+- "--dotformat": Formatting of dot node output.
+- Changed "--noaddr0" to "--noautomaticaddr". Also suppresses the SNA start address.
+- Subroutines are divided into several subroutines if not in a coherent block.
+
 ## 1.1.0
 - Supports output of callgraphs into dot files for visualization with graphviz.
 - Callgraph visualization emphasizes: